Refactor: firebase.js 파일 분리
